refactor(dh): change top bar title service to component store
